refactor(elf): extract indirect-target analysis into a pure tested seam
Lift the ELF indirect-control-flow-target analysis out of ElfPatcher into
a new `elf_patcher::indirect_targets` module exposing a single pure entry
point:

    fn indirect_control_flow_targets(file_data: &[u8]) -> Result<HashSet<u64>, _>

The byte-patcher previously fused two unrelated concerns that shared only
`file_data`: the stateful window validator/patcher, and a stateless scanner
of ELF relocation formats (REL/RELA, compact RELR bitmaps, symbol+addend
resolution, `.rodata`/`.data.rel.ro` pointer scanning). All relocation-format
knowledge — one method plus eight free helpers and the SHT_RELR constant —
now sits behind one small interface; `ElfPatcher::indirect_control_flow_targets`
becomes a one-line delegate, so the public interface and the main.rs caller
are unchanged.

The three soundness tests (ADR-0009 Decision 5) move with the logic and now
exercise the seam directly from crafted ELF bytes, dropping the ElfPatcher
and temp-file setup they previously needed. mod.rs drops 761 net lines.
